+++ src/components/screen/HtmlScrollScreen.jsx
@@ -21,7 +21,7 @@ import BaseScreen, {
 } from './BaseScreen';
 import Connector from '../../service/connector';
 import HtmlContent from '../content/HtmlContent';
-import { FOOTER_INDEX } from '../../constants/CalculationsConstants';
+import { FOOTER_INDEX, PRE_CALCULATION } from '../../constants/CalculationsConstants';
 import DOMEventConstants from '../../constants/DOMEventConstants';
 import { ViewType } from '../../constants/SettingConstants';
 import { getStyledContent, getStyledFooter } from '../styled';
@@ -81,7 +81,7 @@ class HtmlScrollScreen extends BaseScreen {
     return (
       <Footer
         key="footer"
-        isCalculated={footerCalculations.isCalculated}
+        isCalculated={footerCalculations.total !== PRE_CALCULATION}
         content={footer}
         containerVerticalMargin={containerVerticalMargin}
         startOffset={Connector.calculations.getStartOffset(FOOTER_INDEX)}
@@ -94,17 +94,16 @@ class HtmlScrollScreen extends BaseScreen {
     const {
       contentFooter,
     } = this.props;
-    const startOffset = Connector.calculations.getStartOffset(content.index);
+    const { offset, total } = Connector.calculations.getCalculation(content.index);
     const isLastContent = Connector.calculations.isLastContent(content.index);
-    const isCalculated = Connector.calculations.isContentCalculated(content.index);
 
     return (
       <HtmlContent
         key={`${content.uri}:${content.index}`}
         ref={this.getContentRef(content.index)}
         content={content}
-        isCalculated={isCalculated}
-        startOffset={startOffset}
+        isCalculated={total !== PRE_CALCULATION}
+        startOffset={offset}
         contentFooter={isLastContent ? contentFooter : null}
         StyledContent={StyledContent}
       />